LPS-71131 Move Dummy* to the export-import-test-util since it can be useful later
