+++ modules/apps/web-experience/export-import/export-import-test-util/bnd.bnd
@@ -1,6 +1,11 @@
 Bundle-Name: Liferay Export Import Test Utilities
 Bundle-SymbolicName: com.liferay.exportimport.test.util
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.exportimport.test.util
+Export-Package:\
+	com.liferay.exportimport.test.util,\
+	com.liferay.exportimport.test.util.constants,\
+	com.liferay.exportimport.test.util.exportimport.data.handler,\
+	com.liferay.exportimport.test.util.model,\
+	com.liferay.exportimport.test.util.model.util
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Data Management
