+++ modules/apps/web-experience/export-import/export-import-test-util/build.gradle
@@ -1,8 +1,18 @@
+sourceCompatibility = "1.8"
+targetCompatibility = "1.8"
+
 dependencies {
+	provided group: "com.liferay", name: "com.liferay.xstream.configurator.api", version: "2.0.0"
+	provided group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.4.0"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.22.0"
 	provided group: "com.liferay.portal", name: "com.liferay.portal.test", version: "1.0.0"
+	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
 	provided group: "junit", name: "junit", version: "4.12"
+	provided group: "org.hibernate", name: "hibernate-core", version: "3.6.10.Final"
+	provided group: "org.osgi", name: "org.osgi.core", version: "6.0.0"
+	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:web-experience:export-import:export-import-api")
+	provided project(":apps:web-experience:export-import:export-import-service")
 }
 
 liferay {
