+++ modules/apps/web-experience/export-import/export-import-test/src/testIntegration/java/com/liferay/exportimport/test/DummyFolderExportImportTest.java
@@ -18,9 +18,9 @@ import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.exportimport.kernel.lar.PortletDataHandlerKeys;
 import com.liferay.exportimport.staged.model.repository.StagedModelRepository;
 import com.liferay.exportimport.staged.model.repository.StagedModelRepositoryRegistryUtil;
-import com.liferay.exportimport.test.constants.DummyFolderPortletKeys;
-import com.liferay.exportimport.test.model.DummyFolder;
-import com.liferay.exportimport.test.model.util.DummyFolderTestUtil;
+import com.liferay.exportimport.test.util.constants.DummyFolderPortletKeys;
+import com.liferay.exportimport.test.util.model.DummyFolder;
+import com.liferay.exportimport.test.util.model.util.DummyFolderTestUtil;
 import com.liferay.portal.dao.orm.hibernate.DynamicQueryFactoryImpl;
 import com.liferay.portal.kernel.dao.orm.DynamicQueryFactoryUtil;
 import com.liferay.portal.kernel.exception.PortalException;
