+++ modules/apps/web-experience/export-import/export-import-test/src/testIntegration/java/com/liferay/exportimport/test/ExportedMissingReferenceExportImportTest.java
@@ -22,16 +22,16 @@ import com.liferay.exportimport.kernel.lar.PortletDataHandler;
 import com.liferay.exportimport.kernel.lar.PortletDataHandlerKeys;
 import com.liferay.exportimport.staged.model.repository.StagedModelRepository;
 import com.liferay.exportimport.staged.model.repository.StagedModelRepositoryRegistryUtil;
-import com.liferay.exportimport.test.constants.DummyFolderPortletKeys;
-import com.liferay.exportimport.test.exportimport.data.handler.DummyFolderWithMissingDummyPortletDataHandler;
-import com.liferay.exportimport.test.exportimport.data.handler.DummyFolderWithMissingLayoutPortletDataHandler;
-import com.liferay.exportimport.test.model.Dummy;
-import com.liferay.exportimport.test.model.DummyFolder;
-import com.liferay.exportimport.test.model.DummyReference;
-import com.liferay.exportimport.test.model.util.DummyFolderTestUtil;
-import com.liferay.exportimport.test.model.util.DummyReferenceTestUtil;
-import com.liferay.exportimport.test.model.util.DummyTestUtil;
 import com.liferay.exportimport.test.util.TestUserIdStrategy;
+import com.liferay.exportimport.test.util.constants.DummyFolderPortletKeys;
+import com.liferay.exportimport.test.util.exportimport.data.handler.DummyFolderWithMissingDummyPortletDataHandler;
+import com.liferay.exportimport.test.util.exportimport.data.handler.DummyFolderWithMissingLayoutPortletDataHandler;
+import com.liferay.exportimport.test.util.model.Dummy;
+import com.liferay.exportimport.test.util.model.DummyFolder;
+import com.liferay.exportimport.test.util.model.DummyReference;
+import com.liferay.exportimport.test.util.model.util.DummyFolderTestUtil;
+import com.liferay.exportimport.test.util.model.util.DummyReferenceTestUtil;
+import com.liferay.exportimport.test.util.model.util.DummyTestUtil;
 import com.liferay.portal.dao.orm.hibernate.DynamicQueryFactoryImpl;
 import com.liferay.portal.kernel.dao.orm.DynamicQueryFactoryUtil;
 import com.liferay.portal.kernel.portlet.PortletBag;
